+++ src/Console/ValidateSchemaCommand.php
@@ -4,8 +4,6 @@ namespace Nuwave\Lighthouse\Console;
 
 use GraphQL\Type\Schema;
 use Illuminate\Console\Command;
-use Illuminate\Contracts\Cache\Repository as CacheRepository;
-use Illuminate\Contracts\Config\Repository as ConfigRepository;
 use Illuminate\Contracts\Events\Dispatcher as EventsDispatcher;
 use Nuwave\Lighthouse\Events\ValidateSchema;
 use Nuwave\Lighthouse\Schema\DirectiveLocator;
@@ -21,17 +19,13 @@ class ValidateSchemaCommand extends Command
     protected $description = 'Validate the GraphQL schema definition.';
 
     public function handle(
-        CacheRepository $cache,
-        ConfigRepository $config,
         EventsDispatcher $eventsDispatcher,
         SchemaBuilder $schemaBuilder,
         DirectiveLocator $directiveLocator,
         TypeRegistry $typeRegistry
     ): void {
         // Clear the cache so this always validates the current schema
-        $cache->forget(
-            $config->get('lighthouse.cache.key')
-        );
+        $this->call(ClearCacheCommand::NAME);
 
         $originalSchema = $schemaBuilder->schema();
         $schemaConfig = $originalSchema->getConfig();